Former-commit-id: d4e3581d70a74b1eda5236650d16b82774d55ebc

+++ src/Core/BoundedContexts/CAPI/WB.Core.BoundedContexts.CAPI/Views/InterviewDetails/IQuestionnaireItemViewModel.cs
@@ -9,6 +9,6 @@ namespace WB.Core.BoundedContexts.Capi.Views.InterviewDetails
     {
         InterviewItemId PublicKey { get; }
         string Text { get; }
-        IQuestionnaireItemViewModel Clone(int[] propagationVector);
+        IQuestionnaireItemViewModel Clone(decimal[] propagationVector);
     }
 }
