+++ src/Core/SharedKernels/DataCollection/WB.Core.SharedKernels.DataCollection/Events/Interview/MultipleOptionsQuestionAnswered.cs
@@ -7,7 +7,7 @@ namespace WB.Core.SharedKernels.DataCollection.Events.Interview
     {
         public decimal[] SelectedValues { get; private set; }
 
-        public MultipleOptionsQuestionAnswered(Guid userId, Guid questionId, int[] propagationVector, DateTime answerTime, decimal[] selectedValues)
+        public MultipleOptionsQuestionAnswered(Guid userId, Guid questionId, decimal[] propagationVector, DateTime answerTime, decimal[] selectedValues)
             : base(userId, questionId, propagationVector, answerTime)
         {
             this.SelectedValues = selectedValues;
